[refactor] ResultDto 응답 필드를 문자열 타입에서 HttpStatus 열거형으로 변경
