+++ src/main/java/com/MooBoo/MooBoo_Spring/Result.java
@@ -2,15 +2,16 @@ package com.MooBoo.MooBoo_Spring;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import org.springframework.http.HttpStatus;
 
 @Data
 @AllArgsConstructor
 public class Result<T> {
     private String message;
-    private String status;
+    private HttpStatus status;
     private T data;
 
-    public static <T> Result<T> wrapper(String message, String status,T data) {
+    public static <T> Result<T> wrapper(String message, HttpStatus status,T data) {
         Result<T> result = new Result<>(message, status, data);
         return result;
     }
